chore(tile-converter): move ext_feature_metadata types from 3dTiles module
