+++ modules/tile-converter/src/i3s-converter/types.ts
@@ -122,3 +122,54 @@ export type I3SMaterialWithTexture = {
   /** Texture content (image) */
   texture?: ImageDataType;
 };
+
+/**
+ * 3DTilesNext EXT_feature_metadata extension
+ * Spec - https://github.com/CesiumGS/glTF/tree/3d-tiles-next/extensions/2.0/Vendor/EXT_feature_metadata
+ */
+export type ExtFeatureMetadata = {
+  /** Feature ids definition in attributes */
+  featureIdAttributes?: ExtFeatureMetadataAttribute[];
+  /** Feature ids definition in textures */
+  featureIdTextures?: ExtFeatureMetadataAttribute[];
+};
+
+/**
+ * Attribute which described featureIds definition.
+ */
+export type ExtFeatureMetadataAttribute = {
+  /** Name of feature table */
+  featureTable: string;
+  /** Described how feature ids are defined */
+  featureIds: ExtFeatureMetadataFeatureIds;
+};
+
+/**
+ * Defining featureIds by attributes or implicitly.
+ */
+type ExtFeatureMetadataFeatureIds = {
+  /** Name of attribute where featureIds are defined */
+  attribute?: string;
+  /** Sets a constant feature ID for each vertex. The default is 0. */
+  constant?: number;
+  /** Sets the rate at which feature IDs increment.
+   * If divisor is zero then constant is used.
+   * If divisor is greater than zero the feature ID increments once per divisor sets of vertices, starting at constant.
+   * The default is 0
+   */
+  divisor?: number;
+  /** gLTF textureInfo object - https://github.com/CesiumGS/glTF/blob/3d-tiles-next/specification/2.0/schema/textureInfo.schema.json */
+  texture?: ExtFeatureMetadataTexture;
+  /** Must be a single channel ("r", "g", "b", or "a") */
+  channels?: 'r' | 'g' | 'b' | 'a';
+};
+
+/**
+ * Reference to a texture.
+ */
+type ExtFeatureMetadataTexture = {
+  /** The set index of texture's TEXCOORD attribute used for texture coordinate mapping.*/
+  texCoord: number;
+  /** The index of the texture. */
+  index: number;
+};
